Cleanup storeprooflib contract, add mock contract for storeprooflib to make testing easier and also add test for mock contract (#59)
* cleanup storeprooflib contract, add mock contract for storeprooflib to make testing easier and also add test for mock contract

* v0.0.0

* add more descriptive commnets and message when revert

* remove redundant abi.encodePacked

* fix tree spacing

* change input/output of function innerHash,leafHash,calculateBlockHash from bytes to bytes32 (more efficient) and also add more desciptive comments

* make contract more consistent by changing hex'20' to uint8(32)

* change uint to uint256 for more consistency

+++ bridges/evm/contracts/ValidatorsManager.sol
@@ -18,7 +18,7 @@ contract ValidatorsManager is Ownable {
   }
 
   function addValidators(address[] memory _validators) public onlyOwner {
-    for (uint i = 0; i < _validators.length; i++) {
+    for (uint256 i = 0; i < _validators.length; i++) {
       require(!validators[_validators[i]], "ALREADY_BE_VALIDATOR");
       validators[_validators[i]] = true;
     }
@@ -26,7 +26,7 @@ contract ValidatorsManager is Ownable {
   }
 
   function removeValidators(address[] memory _validators) public onlyOwner {
-    for (uint i = 0; i < _validators.length; i++) {
+    for (uint256 i = 0; i < _validators.length; i++) {
       require(validators[_validators[i]], "UNKNOWN_VALIDATOR");
       validators[_validators[i]] = false;
     }
